Changes to accommodate ordered factors
When dealing with ordered factors, the class function outputs '"ordered" "factor"', meaning that all of the if statements on lines 44-47 are FALSE and that a non-numeric variable is passed on for the correlation analysis, resulting in an error. The changes here allow the omixerCorr function to handle ordered factors without resulting in an error. Additionally, since the output of chisq.test doesn't have a "statistics" component, the cv variable will always end up being 0. The minor change on line 52 results in the statistic being correctly stored.